goods_id类型由long改为int

+++ src/main/java/com/lebo/entity/GiveGoods.java
@@ -11,7 +11,7 @@ public class GiveGoods {
     private Long id;
     private String fromUserId;
     private String toUserId;
-    private Long goodsId;
+    private Integer goodsId;
     private Integer quantity;
     private String postId;
     private Date giveDate;
@@ -40,11 +40,11 @@ public class GiveGoods {
         this.toUserId = toUserId;
     }
 
-    public Long getGoodsId() {
+    public Integer getGoodsId() {
         return goodsId;
     }
 
-    public void setGoodsId(Long goodsId) {
+    public void setGoodsId(Integer goodsId) {
         this.goodsId = goodsId;
     }
